All TablatureBlock labels are a text area, effectively elinating the need for a NoteBlock. RIP
